Make decorator dir in tests and project
